LogFactoryService class for standard Commons Logging service discovery
Issue: SPR-16585
Issue: SPR-15903
