+++ spring-jcl/src/main/resources/META-INF/services/org.apache.commons.logging.LogFactory
@@ -1,0 +1,1 @@
+org.apache.commons.logging.LogFactoryService
